Separate out integration tests from one giant class into multiple
